edit_workspace_location

+++ database/migrations/2018_11_29_132503_remove_trainer_id_column_in_users_table.php
@@ -13,11 +13,13 @@ class RemoveTrainerIdColumnInUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('trainer_id');
-            $table->dropColumn('seat_id');
-            $table->dropColumn('images');
-        });
+        if (Schema::hasColumn('trainer_id', 'seat_id', 'images')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('trainer_id');
+                $table->dropColumn('seat_id');
+                $table->dropColumn('images');
+            });
+        }
     }
 
     /**
